pkp/pkp-lib#10292 refactored interest manager to repo and removed it
