+++ classes/user/interest/Repository.php
@@ -16,6 +16,7 @@ namespace PKP\user\interest;
 
 use APP\facades\Repo;
 use Illuminate\Support\Facades\DB;
+use PKP\user\User;
 use PKP\user\interest\UserInterest;
 use PKP\controlledVocab\ControlledVocabEntry;
 use Throwable;
@@ -23,10 +24,66 @@ use Throwable;
 class Repository
 {
     /**
-     * Update a user's set of interests
+     * Get all interests for all users in the system
      */
-    public function setUserInterests(array $interests, int $userId): void
+    public function getAllInterests(?string $filter = null): array
     {
+        $controlledVocab = Repo::controlledVocab()->build(
+            UserInterest::CONTROLLED_VOCAB_INTEREST
+        );
+
+        return ControlledVocabEntry::query()
+            ->withControlledVocabId($controlledVocab->id)
+            ->when(
+                $filter,
+                fn($query) => $query->withSetting(
+                    UserInterest::CONTROLLED_VOCAB_INTEREST,
+                    $filter
+                )
+            )
+            ->get()
+            ->sortBy(UserInterest::CONTROLLED_VOCAB_INTEREST)
+            ->pluck(UserInterest::CONTROLLED_VOCAB_INTEREST)
+            ->toArray();
+    }
+
+    /**
+     * Get user reviewing interests. (Cached in memory for batch fetches.)
+     */
+    public function getInterestsForUser(User $user): array
+    {
+        return ControlledVocabEntry::query()
+            ->whereHas(
+                "controlledVocab",
+                fn($query) => $query
+                    ->withSymbolic(UserInterest::CONTROLLED_VOCAB_INTEREST)
+                    ->withAssoc(0, 0)
+            )
+            ->whereHas("userInterest", fn($query) => $query->withUserId($user->getId()))
+            ->get()
+            ->pluck(UserInterest::CONTROLLED_VOCAB_INTEREST, 'id')
+            ->toArray();
+    }
+
+    /**
+     * Returns a comma separated string of a user's interests
+     */
+    public function getInterestsString(User $user): string
+    {
+        $interests = $this->getInterestsForUser($user);
+
+        return implode(', ', $interests);
+    }
+
+    /**
+     * Set a user's interests
+     */
+    public function setInterestsForUser(User $user, string|array|null $interests = null): void
+    {
+        $interests = is_array($interests)
+            ? $interests
+            : (empty($interests) ? [] : explode(',', $interests));
+
         $controlledVocab = Repo::controlledVocab()->build(
             UserInterest::CONTROLLED_VOCAB_INTEREST
         );
@@ -47,7 +104,7 @@ class Repository
             DB::beginTransaction();
 
             // Delete the existing interests association.
-            UserInterest::query()->withUserId($userId)->delete();
+            UserInterest::query()->withUserId($user->getId())->delete();
             
             $newInterestIds = collect(
                     array_diff(
@@ -70,7 +127,7 @@ class Repository
             collect($currentInterests->pluck('id'))
                 ->merge($newInterestIds)
                 ->each(fn ($interestId) => UserInterest::create([
-                    'userId' => $userId,
+                    'userId' => $user->getId(),
                     'controlledVocabEntryId' => $interestId,
                 ]));
